+++ examples/user_info_custom_provider/src/main.rs
@@ -1,9 +1,7 @@
-#![feature(decl_macro, proc_macro_hygiene)]
-
-use rocket::http::{Cookie, Cookies};
+use rocket::http::{Cookie, CookieJar};
 use rocket::request::{self, FromRequest, Request};
 use rocket::response::Redirect;
-use rocket::{get, routes, Outcome};
+use rocket::{get, routes};
 
 mod oauth2;
 
@@ -11,18 +9,19 @@ struct User {
     pub username: String,
 }
 
+#[rocket::async_trait]
 impl<'a, 'r> FromRequest<'a, 'r> for User {
     type Error = ();
 
-    fn from_request(request: &'a Request<'r>) -> request::Outcome<User, ()> {
-        let mut cookies = request.guard::<Cookies<'_>>().expect("request cookies");
+    async fn from_request(request: &'a Request<'r>) -> request::Outcome<User, ()> {
+        let cookies = request.guard::<&CookieJar<'_>>().await.expect("request cookies");
         if let Some(cookie) = cookies.get_private("username") {
-            return Outcome::Success(User {
+            return request::Outcome::Success(User {
                 username: cookie.value().to_string(),
             });
         }
 
-        Outcome::Forward(())
+        request::Outcome::Forward(())
     }
 }
 
@@ -37,14 +36,14 @@ fn index_anonymous() -> &'static str {
 }
 
 #[get("/logout")]
-fn logout(mut cookies: Cookies<'_>) -> Redirect {
+fn logout(cookies: &CookieJar<'_>) -> Redirect {
     cookies.remove(Cookie::named("username"));
     Redirect::to("/")
 }
 
-fn main() {
+#[rocket::launch]
+fn rocket() -> _ {
     rocket::ignite()
         .mount("/", routes![index, index_anonymous, logout])
         .attach(oauth2::github::fairing())
-        .launch();
 }