+++ src/hyper_rustls_adapter.rs
@@ -1,10 +1,9 @@
 use std::convert::TryInto;
-use std::io::Read;
 
 use hyper::{
-    header::{Accept, Authorization, Basic, ContentType},
-    net::HttpsConnector,
-    Client,
+    body::HttpBody,
+    header::{ACCEPT, AUTHORIZATION, CONTENT_TYPE},
+    Body, Client, Request,
 };
 use rocket::http::ext::IntoOwned;
 use rocket::http::uri::Absolute;
@@ -19,21 +18,24 @@ use super::{Adapter, Error, ErrorKind, OAuthConfig, TokenRequest, TokenResponse}
 /// not supported by your authorization server, the [`basic_auth`] method
 /// can be used to change this behavior.
 ///
-/// [`basic_auth`]: HyperSyncRustlsAdapter::basic_auth()
+/// [`basic_auth`]: HyperRustlsAdapter::basic_auth()
+
 #[derive(Clone, Debug)]
-pub struct HyperSyncRustlsAdapter {
+pub struct HyperRustlsAdapter {
     use_basic_auth: bool,
+    client: Client<hyper_rustls::HttpsConnector<hyper::client::HttpConnector>>,
 }
 
-impl Default for HyperSyncRustlsAdapter {
+impl Default for HyperRustlsAdapter {
     fn default() -> Self {
         Self {
             use_basic_auth: true,
+            client: Client::builder().build(hyper_rustls::HttpsConnector::new()),
         }
     }
 }
 
-impl HyperSyncRustlsAdapter {
+impl HyperRustlsAdapter {
     /// Sets whether or not this adapter will use HTTP Basic Authentication.
     /// Although servers are required to support it (RFC 6749 §2.3.1), not all
     /// do.
@@ -45,33 +47,34 @@ impl HyperSyncRustlsAdapter {
     ///
     /// ```rust,no_run
     /// use rocket::fairing::AdHoc;
-    /// use rocket_oauth2::{HyperSyncRustlsAdapter, OAuth2, OAuthConfig, StaticProvider};
+    /// use rocket_oauth2::{HyperRustlsAdapter, OAuth2, OAuthConfig, StaticProvider};
     ///
     /// struct MyProvider;
     ///
-    /// fn main() {
+    /// #[rocket::launch]
+    /// fn rocket() -> _ {
     ///     rocket::ignite()
-    ///         .attach(AdHoc::on_attach("OAuth Config", |rocket| {
-    ///             let config = OAuthConfig::from_config(rocket.config(), "my_provider").unwrap();
+    ///         .attach(AdHoc::on_attach("OAuth Config", |mut rocket| async {
+    ///             let config = OAuthConfig::from_figment(rocket.figment(), "my_provider").unwrap();
     ///             Ok(rocket.attach(OAuth2::<MyProvider>::custom(
-    ///                 HyperSyncRustlsAdapter::default().basic_auth(false), config)
+    ///                 HyperRustlsAdapter::default().basic_auth(false), config)
     ///             ))
     ///         }))
-    ///         .launch();
     /// }
     /// ```
     pub fn basic_auth(self, use_basic_auth: bool) -> Self {
-        Self { use_basic_auth }
+        Self { use_basic_auth, ..self }
     }
 }
 
-impl Adapter for HyperSyncRustlsAdapter {
+#[async_trait::async_trait]
+impl Adapter for HyperRustlsAdapter {
     fn authorization_uri(
         &self,
         config: &OAuthConfig,
         state: &str,
         scopes: &[&str],
-        extras: &[(&str, &str)],
+        extra_params: &[(&str, &str)],
     ) -> Result<Absolute<'static>, Error> {
         let auth_uri = config.provider().auth_uri();
 
@@ -95,7 +98,7 @@ impl Adapter for HyperSyncRustlsAdapter {
 
         // Request parameters must not be included more than once. This
         // adapter chooses to ignore duplicates instead of overwriting.
-        for (name, value) in extras {
+        for (name, value) in extra_params {
             match *name {
                 "response_type" | "client_id" | "state" => continue,
                 "redirect_uri" if config.redirect_uri().is_some() => continue,
@@ -109,59 +112,69 @@ impl Adapter for HyperSyncRustlsAdapter {
             .into_owned())
     }
 
-    fn exchange_code(
+    async fn exchange_code(
         &self,
         config: &OAuthConfig,
         token: TokenRequest,
     ) -> Result<TokenResponse<()>, Error> {
-        let https = HttpsConnector::new(hyper_sync_rustls::TlsClient::new());
-        let client = Client::with_connector(https);
+        let mut request = Request::post(&*config.provider().token_uri())
+            .header(ACCEPT, "application/json")
+            .header(CONTENT_TYPE, "application/x-www-form-urlencoded");
 
-        let mut request = client
-            .post(config.provider().token_uri().as_ref())
-            .header(Accept::json())
-            .header(ContentType::form_url_encoded());
+        let req_str = {
+            let mut ser = UrlSerializer::new(String::new());
+            match token {
+                TokenRequest::AuthorizationCode(code) => {
+                    ser.append_pair("grant_type", "authorization_code");
+                    ser.append_pair("code", &code);
+                    if let Some(redirect_uri) = config.redirect_uri() {
+                        ser.append_pair("redirect_uri", redirect_uri);
+                    }
 
-        let mut ser = UrlSerializer::new(String::new());
-        match token {
-            TokenRequest::AuthorizationCode(code) => {
-                ser.append_pair("grant_type", "authorization_code");
-                ser.append_pair("code", &code);
-                if let Some(redirect_uri) = config.redirect_uri() {
-                    ser.append_pair("redirect_uri", redirect_uri);
+                }
+                TokenRequest::RefreshToken(token) => {
+                    ser.append_pair("grant_type", "refresh_token");
+                    ser.append_pair("refresh_token", &token);
                 }
             }
-            TokenRequest::RefreshToken(token) => {
-                ser.append_pair("grant_type", "refresh_token");
-                ser.append_pair("refresh_token", &token);
+
+            if self.use_basic_auth {
+                let encoded = base64::encode(format!("{}:{}", config.client_id(), config.client_secret()));
+                request = request.header(AUTHORIZATION, format!("Basic {}", encoded))
+            } else {
+                ser.append_pair("client_id", config.client_id());
+                ser.append_pair("client_secret", config.client_secret());
             }
-        }
 
-        if self.use_basic_auth {
-            request = request.header(Authorization(Basic {
-                username: config.client_id().to_string(),
-                password: Some(config.client_secret().to_string()),
-            }));
-        } else {
-            ser.append_pair("client_id", config.client_id());
-            ser.append_pair("client_secret", config.client_secret());
-        }
+            ser.finish()
+        };
 
-        let req_str = ser.finish();
-
-        let request = request.body(&req_str);
-
-        let response = request
-            .send()
+        let request = request
+            .body(Body::from(req_str))
             .map_err(|e| Error::new_from(ErrorKind::ExchangeFailure, e))?;
 
-        if !response.status.is_success() {
+        let response = self
+            .client
+            .request(request)
+            .await
+            .map_err(|e| Error::new_from(ErrorKind::ExchangeFailure, e))?;
+        if !response.status().is_success() {
             return Err(Error::new(ErrorKind::ExchangeError(
-                response.status.to_u16(),
+                response.status().as_u16(),
             )));
         }
 
-        let data: serde_json::Value = serde_json::from_reader(response.take(2 * 1024 * 1024))
+        let mut body = response.into_body();
+        let mut bytes = vec![];
+        while let Some(chunk) = body.data().await {
+            let chunk = chunk.map_err(|e| Error::new_from(ErrorKind::ExchangeFailure, e))?;
+            if bytes.len() + chunk.len() > 2 * 1024 * 1024 {
+                return Err(Error::new_from(ErrorKind::ExchangeFailure, "Response body was too large."));
+            }
+            bytes.extend(chunk);
+        }
+
+        let data: serde_json::Value = serde_json::from_slice(&bytes)
             .map_err(|e| Error::new_from(ErrorKind::ExchangeFailure, e))?;
         Ok(data.try_into()?)
     }